+++ src/roq/huobi_futures/market_data.cpp
@@ -16,13 +16,13 @@
 
 #include "roq/huobi_futures/flags.h"
 
-using namespace roq::literals;
+using namespace std::literals;
 
 namespace roq {
 namespace huobi_futures {
 
 namespace {
-static const auto NAME = "md"_sv;
+static const auto NAME = "md"sv;
 static const auto SUPPORTS = utils::Mask{
     SupportType::TOP_OF_BOOK,
     SupportType::MARKET_BY_PRICE,
@@ -49,7 +49,7 @@ void emplace(MBPUpdate &result, const T &value) {
 
 MarketData::MarketData(
     Handler &handler, core::io::Context &context, uint32_t stream_id, Shared &shared)
-    : handler_(handler), stream_id_(stream_id), name_(fmt::format("{}:{}"_sv, stream_id_, NAME)),
+    : handler_(handler), stream_id_(stream_id), name_(fmt::format("{}:{}"sv, stream_id_, NAME)),
       connection_(
           *this,
           context,
@@ -62,22 +62,22 @@ MarketData::MarketData(
       decode_buffer_(Flags::decode_buffer_size()),
       request_id_(static_cast<uint64_t>(stream_id_) * 1000000),  // scale (debugging)
       counter_{
-          .disconnect = create_metrics(name_, "disconnect"_sv),
+          .disconnect = create_metrics(name_, "disconnect"sv),
       },
       profile_{
-          .parse = create_metrics(name_, "parse"_sv),
-          .error = create_metrics(name_, "error"_sv),
-          .result = create_metrics(name_, "result"_sv),
-          .agg_trade = create_metrics(name_, "agg_trade"_sv),
-          .trade = create_metrics(name_, "trade"_sv),
-          .mini_ticker = create_metrics(name_, "mini_ticker"_sv),
-          .book_ticker = create_metrics(name_, "book_ticker"_sv),
-          .depth = create_metrics(name_, "depth"_sv),
-          .depth_update = create_metrics(name_, "depth_update"_sv),
+          .parse = create_metrics(name_, "parse"sv),
+          .error = create_metrics(name_, "error"sv),
+          .result = create_metrics(name_, "result"sv),
+          .agg_trade = create_metrics(name_, "agg_trade"sv),
+          .trade = create_metrics(name_, "trade"sv),
+          .mini_ticker = create_metrics(name_, "mini_ticker"sv),
+          .book_ticker = create_metrics(name_, "book_ticker"sv),
+          .depth = create_metrics(name_, "depth"sv),
+          .depth_update = create_metrics(name_, "depth_update"sv),
       },
       latency_{
-          .ping = create_metrics(name_, "ping"_sv),
-          .heartbeat = create_metrics(name_, "heartbeat"_sv),
+          .ping = create_metrics(name_, "ping"sv),
+          .heartbeat = create_metrics(name_, "heartbeat"sv),
       },
       shared_(shared), download_({}, [this](auto state) { return download(state); }) {
 }
@@ -170,7 +170,7 @@ void MarketData::operator()(const core::web::Socket::Text &text) {
 }
 
 void MarketData::operator()(const core::web::Socket::Binary &) {
-  log::fatal("Unexpected"_sv);
+  log::fatal("Unexpected"sv);
 }
 
 void MarketData::operator()(ConnectionStatus status) {
@@ -184,7 +184,7 @@ void MarketData::operator()(ConnectionStatus status) {
         .type = StreamType::WEB_SOCKET,
         .priority = Priority::PRIMARY,
     };
-    log::info("stream_status={}"_sv, stream_status);
+    log::info("stream_status={}"sv, stream_status);
     server::create_trace_and_dispatch(handler_, trace_info, stream_status);
   }
 }
@@ -226,8 +226,8 @@ void MarketData::subscribe_agg_trade(const roq::span<std::string> &symbols) {
       R"("method":"SUBSCRIBE",)"
       R"("params":["{}@aggTrade"],)"
       R"("id":{})"
-      R"(}})"_sv,
-      fmt::join(symbols, R"(@aggTrade",")"_sv),
+      R"(}})"sv,
+      fmt::join(symbols, R"(@aggTrade",")"sv),
       id);
   connection_.send_text(message);
 }
@@ -240,8 +240,8 @@ void MarketData::subscribe_trade(const roq::span<std::string> &symbols) {
       R"("method":"SUBSCRIBE",)"
       R"("params":["{}@trade"],)"
       R"("id":{})"
-      R"(}})"_sv,
-      fmt::join(symbols, R"(@trade",")"_sv),
+      R"(}})"sv,
+      fmt::join(symbols, R"(@trade",")"sv),
       id);
   connection_.send_text(message);
 }
@@ -254,8 +254,8 @@ void MarketData::subscribe_mini_ticker(const roq::span<std::string> &symbols) {
       R"("method":"SUBSCRIBE",)"
       R"("params":["{}@miniTicker"],)"
       R"("id":{})"
-      R"(}})"_sv,
-      fmt::join(symbols, R"(@miniTicker",")"_sv),
+      R"(}})"sv,
+      fmt::join(symbols, R"(@miniTicker",")"sv),
       id);
   connection_.send_text(message);
 }
@@ -268,8 +268,8 @@ void MarketData::subscribe_book_ticker(const roq::span<std::string> &symbols) {
       R"("method":"SUBSCRIBE",)"
       R"("params":["{}@bookTicker"],)"
       R"("id":{})"
-      R"(}})"_sv,
-      fmt::join(symbols, R"(@bookTicker",")"_sv),
+      R"(}})"sv,
+      fmt::join(symbols, R"(@bookTicker",")"sv),
       id);
   connection_.send_text(message);
 }
@@ -277,18 +277,18 @@ void MarketData::subscribe_book_ticker(const roq::span<std::string> &symbols) {
 void MarketData::subscribe_depth(const roq::span<std::string> &symbols) {
   assert(!symbols.empty());
   auto stream = fmt::format(
-      R"(@depth{}@{}ms)"_sv,
+      R"(@depth{}@{}ms)"sv,
       Flags::ws_subscribe_depth_levels(),
       std::chrono::duration_cast<std::chrono::milliseconds>(Flags::ws_subscribe_depth_freq())
           .count());
   auto id = ++request_id_;
-  auto separator = fmt::format(R"({}",")"_sv, stream);
+  auto separator = fmt::format(R"({}",")"sv, stream);
   auto message = fmt::format(
       R"({{)"
       R"("method":"SUBSCRIBE",)"
       R"("params":["{}{}"],)"
       R"("id":{})"
-      R"(}})"_sv,
+      R"(}})"sv,
       fmt::join(symbols, separator),
       stream,
       id);
@@ -302,23 +302,23 @@ void MarketData::parse(const std::string_view &message) {
       core::json::Buffer buffer(decode_buffer_);
       json::MarketStreamParser::dispatch(*this, message, buffer, trace_info);
     } catch (...) {
-      log::warn(R"(message="{}")"_sv, message);
+      log::warn(R"(message="{}")"sv, message);
       core::tools::UnhandledException::terminate();
     }
   });
 }
 
 void MarketData::operator()(int32_t id, const json::Error &error) {
-  profile_.error([&]() { log::warn("id={}, error={}"_sv, id, error); });
+  profile_.error([&]() { log::warn("id={}, error={}"sv, id, error); });
 }
 
 void MarketData::operator()(int32_t id, const json::Result &result) {
-  profile_.result([&]() { log::info("id={}, result={}"_sv, id, result); });
+  profile_.result([&]() { log::info("id={}, result={}"sv, id, result); });
 }
 
 void MarketData::operator()(const json::AggTrade &agg_trade, const server::TraceInfo &trace_info) {
   profile_.agg_trade([&]() {
-    log::info<3>("agg_trade={}"_sv, agg_trade);
+    log::info<3>("agg_trade={}"sv, agg_trade);
     auto side = agg_trade.buyer_is_maker ? Side::BUY : Side::SELL;
     Trade trade{
         .side = side,
@@ -340,7 +340,7 @@ void MarketData::operator()(const json::AggTrade &agg_trade, const server::Trace
 
 void MarketData::operator()(const json::Trade &trade, const server::TraceInfo &trace_info) {
   profile_.trade([&]() {
-    log::info<3>("trade={}"_sv, trade);
+    log::info<3>("trade={}"sv, trade);
     auto side = trade.buyer_is_maker ? Side::BUY : Side::SELL;
     Trade trade_{
         .side = side,
@@ -363,7 +363,7 @@ void MarketData::operator()(const json::Trade &trade, const server::TraceInfo &t
 void MarketData::operator()(
     const json::MiniTicker &mini_ticker, const server::TraceInfo &trace_info) {
   profile_.mini_ticker([&]() {
-    log::info<3>("mini_ticker={}"_sv, mini_ticker);
+    log::info<3>("mini_ticker={}"sv, mini_ticker);
     Statistics statistics[] = {
         {.type = StatisticsType::HIGHEST_TRADED_PRICE, .value = mini_ticker.high_price},
         {.type = StatisticsType::LOWEST_TRADED_PRICE, .value = mini_ticker.low_price},
@@ -385,7 +385,7 @@ void MarketData::operator()(
 void MarketData::operator()(
     const json::BookTicker &book_ticker, const server::TraceInfo &trace_info) {
   profile_.book_ticker([&]() {
-    log::info<3>("book_ticker={}"_sv, book_ticker);
+    log::info<3>("book_ticker={}"sv, book_ticker);
     TopOfBook top_of_book{
         .stream_id = stream_id_,
         .exchange = Flags::exchange(),
@@ -406,7 +406,7 @@ void MarketData::operator()(
 void MarketData::operator()(
     const std::string_view &symbol, const json::Depth &depth, const server::TraceInfo &trace_info) {
   profile_.depth([&]() {
-    log::info<3>(R"(symbol="{}", depth={})"_sv, symbol, depth);
+    log::info<3>(R"(symbol="{}", depth={})"sv, symbol, depth);
     core::back_emplacer bids(shared_.bids), asks(shared_.asks);
     for (auto &item : depth.bids)
       bids.emplace_back([&item](auto &result) { emplace(result, item); });
@@ -432,7 +432,7 @@ void MarketData::operator()(
     const json::DepthUpdate &depth_update,
     const server::TraceInfo &) {
   profile_.depth_update([&]() {
-    log::info<3>(R"(symbol="{}", depth_update={})"_sv, symbol, depth_update);
+    log::info<3>(R"(symbol="{}", depth_update={})"sv, symbol, depth_update);
     // do nothing
     // XXX why?
   });