+++ src/roq/huobi_futures/order_entry.cpp
@@ -13,13 +13,13 @@
 
 #include "roq/huobi_futures/json/utils.h"
 
-using namespace roq::literals;
+using namespace std::literals;
 
 namespace roq {
 namespace huobi_futures {
 
 namespace {
-static const auto NAME = "om"_sv;
+static const auto NAME = "om"sv;
 static const auto SUPPORTS = utils::Mask{
     SupportType::REFERENCE_DATA,
     SupportType::MARKET_STATUS,
@@ -44,7 +44,7 @@ OrderEntry::OrderEntry(
     Security &security,
     Shared &shared)
     : handler_(handler), stream_id_(stream_id),
-      name_(fmt::format("{}:{}:{}"_sv, stream_id_, NAME, security.get_account())),
+      name_(fmt::format("{}:{}:{}"sv, stream_id_, NAME, security.get_account())),
       connection_(
           *this,
           context,
@@ -61,22 +61,22 @@ OrderEntry::OrderEntry(
           Flags::rest_ping_path()),
       decode_buffer_(Flags::decode_buffer_size()),
       counter_{
-          .disconnect = create_metrics(name_, "disconnect"_sv),
+          .disconnect = create_metrics(name_, "disconnect"sv),
       },
       profile_{
-          .listen_key = create_metrics(name_, "listen_key"_sv),
-          .listen_key_ack = create_metrics(name_, "listen_key_ack"_sv),
-          .account = create_metrics(name_, "account"_sv),
-          .account_ack = create_metrics(name_, "account_ack"_sv),
-          .exchange_info = create_metrics(name_, "exchange_info"_sv),
-          .exchange_info_ack = create_metrics(name_, "exchange_info_ack"_sv),
-          .new_order = create_metrics(name_, "new_order"_sv),
-          .new_order_ack = create_metrics(name_, "new_order_ack"_sv),
-          .cancel_order = create_metrics(name_, "cancel_order"_sv),
-          .cancel_order_ack = create_metrics(name_, "cancel_order_ack"_sv),
+          .listen_key = create_metrics(name_, "listen_key"sv),
+          .listen_key_ack = create_metrics(name_, "listen_key_ack"sv),
+          .account = create_metrics(name_, "account"sv),
+          .account_ack = create_metrics(name_, "account_ack"sv),
+          .exchange_info = create_metrics(name_, "exchange_info"sv),
+          .exchange_info_ack = create_metrics(name_, "exchange_info_ack"sv),
+          .new_order = create_metrics(name_, "new_order"sv),
+          .new_order_ack = create_metrics(name_, "new_order_ack"sv),
+          .cancel_order = create_metrics(name_, "cancel_order"sv),
+          .cancel_order_ack = create_metrics(name_, "cancel_order_ack"sv),
       },
       latency_{
-          .ping = create_metrics(name_, "ping"_sv),
+          .ping = create_metrics(name_, "ping"sv),
       },
       security_(security), shared_(shared),
       download_(Flags::rest_request_timeout(), [this](auto state) { return download(state); }) {
@@ -139,7 +139,7 @@ uint16_t OrderEntry::operator()(
 
 uint16_t OrderEntry::operator()(
     const Event<CancelAllOrders> &, [[maybe_unused]] const std::string_view &request_id) {
-  log::fatal("*** CANCEL ALL ORDERS *NOT* SUPPORTED ***"_sv);
+  log::fatal("*** CANCEL ALL ORDERS *NOT* SUPPORTED ***"sv);
 }
 
 void OrderEntry::operator()(const core::web::Client::Connected &) {
@@ -179,7 +179,7 @@ void OrderEntry::operator()(ConnectionStatus status) {
         .type = StreamType::REST,
         .priority = Priority::PRIMARY,
     };
-    log::info("stream_status={}"_sv, stream_status);
+    log::info("stream_status={}"sv, stream_status);
     server::create_trace_and_dispatch(handler_, trace_info, stream_status);
   }
 }
@@ -211,8 +211,8 @@ uint32_t OrderEntry::download(OrderEntryState state) {
 void OrderEntry::get_listen_key() {
   profile_.listen_key([&]() {
     auto method = core::http::Method::POST;
-    auto path = "/api/v3/userDataStream"_sv;
-    auto headers = fmt::format("X-MBX-APIKEY: {}\r\n"_sv, security_.get_api_key());
+    auto path = "/api/v3/userDataStream"sv;
+    auto headers = fmt::format("X-MBX-APIKEY: {}\r\n"sv, security_.get_api_key());
     core::web::Request request{
         .method = method,
         .path = path,
@@ -226,7 +226,7 @@ void OrderEntry::get_listen_key() {
     };
     auto sequence = download_.sequence();
     connection_(
-        "listen_key"_sv,
+        "listen_key"sv,
         request,
         [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
           auto trace_info = server::create_trace_info();
@@ -243,9 +243,9 @@ void OrderEntry::get_listen_key_ack(
     auto state = OrderEntryState::LISTEN_KEY;
     try {
       auto [status, category, body] = response.result();
-      log::debug(R"(status={}, category={}, body="{}")"_sv, status, category, body);
+      log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
       if (download_.skip(sequence, state)) {
-        log::info("Download state={} has already been processed"_sv, state);
+        log::info("Download state={} has already been processed"sv, state);
         return;
       }
       response.expect(core::http::Status::OK);
@@ -254,7 +254,7 @@ void OrderEntry::get_listen_key_ack(
       (*this)(event);
       download_.check(state);
     } catch (core::NetworkError &e) {
-      log::warn(R"(Exception type={}, what="{}")"_sv, typeid(e).name(), e.what());
+      log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
       download_.retry(state);
     }
   });
@@ -262,11 +262,11 @@ void OrderEntry::get_listen_key_ack(
 
 void OrderEntry::operator()(const server::Trace<json::ListenKey> &event) {
   auto &[trace_info, listen_key] = event;
-  log::info<2>("listen_key={}"_sv, listen_key);
+  log::info<2>("listen_key={}"sv, listen_key);
   bool initial = listen_key_.empty();
   if (utils::update(listen_key_, listen_key.listen_key)) {
     if (initial) {
-      log::info(R"(Listen key has been acquired (value="{}"))"_sv, listen_key_);
+      log::info(R"(Listen key has been acquired (value="{}"))"sv, listen_key_);
       ListenKeyUpdate listen_key_update{
           .account = security_.get_account(),
           .listen_key = listen_key.listen_key,
@@ -274,7 +274,7 @@ void OrderEntry::operator()(const server::Trace<json::ListenKey> &event) {
       create_trace_and_dispatch(handler_, trace_info, listen_key_update);
     } else {
       if (ROQ_UNLIKELY(!initial))
-        log::info("Listen key has been refreshed!"_sv);
+        log::info("Listen key has been refreshed!"sv);
     }
   }
   auto now = core::get_system_clock();
@@ -287,7 +287,7 @@ void OrderEntry::refresh_listen_key() {
   auto now = core::get_system_clock();
   if (listen_key_refresh_ == listen_key_refresh_.zero() || now < listen_key_refresh_)
     return;
-  log::info("Refreshing listen key..."_sv);
+  log::info("Refreshing listen key..."sv);
   listen_key_refresh_ = now + Flags::rest_listen_key_refresh();
   get_listen_key();
   /*
@@ -295,7 +295,7 @@ void OrderEntry::refresh_listen_key() {
     try {
       (*this)(promise.get());
     } catch (core::NetworkError &) {
-      log::warn("Rescheduling listen key refresh!"_sv);
+      log::warn("Rescheduling listen key refresh!"sv);
       auto now = core::get_system_clock();
       listen_key_refresh_ = now + Flags::rest_listen_key_refresh();
     }
@@ -308,11 +308,11 @@ void OrderEntry::refresh_listen_key() {
 void OrderEntry::get_account() {
   profile_.account([&]() {
     auto method = core::http::Method::GET;
-    auto path = "/api/v3/account"_sv;
+    auto path = "/api/v3/account"sv;
     auto now = core::get_realtime_clock();
     auto [timestamp, signature] = security_.create_signature(now);
-    auto query = fmt::format("?{}&signature={}"_sv, timestamp, signature);
-    auto headers = fmt::format("X-MBX-APIKEY: {}\r\n"_sv, security_.get_api_key());
+    auto query = fmt::format("?{}&signature={}"sv, timestamp, signature);
+    auto headers = fmt::format("X-MBX-APIKEY: {}\r\n"sv, security_.get_api_key());
     core::web::Request request{
         .method = method,
         .path = path,
@@ -326,7 +326,7 @@ void OrderEntry::get_account() {
     };
     auto sequence = download_.sequence();
     connection_(
-        "account"_sv, request, [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
+        "account"sv, request, [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
           auto trace_info = server::create_trace_info();
           server::Trace event(trace_info, response);
           get_account_ack(event, sequence);
@@ -341,9 +341,9 @@ void OrderEntry::get_account_ack(
     auto state = OrderEntryState::ACCOUNT;
     try {
       auto [status, category, body] = response.result();
-      log::debug(R"(status={}, category={}, body="{}")"_sv, status, category, body);
+      log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
       if (download_.skip(sequence, state)) {
-        log::info("Download state={} has already been processed"_sv, state);
+        log::info("Download state={} has already been processed"sv, state);
         return;
       }
       response.expect(core::http::Status::OK);
@@ -353,7 +353,7 @@ void OrderEntry::get_account_ack(
       (*this)(event);
       download_.check(state);
     } catch (core::NetworkError &e) {
-      log::warn(R"(Exception type={}, what="{}")"_sv, typeid(e).name(), e.what());
+      log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
       download_.retry(state);
     }
   });
@@ -361,7 +361,7 @@ void OrderEntry::get_account_ack(
 
 void OrderEntry::operator()(const server::Trace<json::Account> &event) {
   auto &[trace_info, account] = event;
-  log::info<1>("account={}"_sv, account);
+  log::info<1>("account={}"sv, account);
   for (auto &item : account.balances) {
     FundsUpdate funds_update{
         .stream_id = stream_id_,
@@ -380,7 +380,7 @@ void OrderEntry::operator()(const server::Trace<json::Account> &event) {
 void OrderEntry::get_exchange_info() {
   profile_.exchange_info([&]() {
     auto method = core::http::Method::GET;
-    auto path = "/api/v3/exchangeInfo"_sv;
+    auto path = "/api/v3/exchangeInfo"sv;
     core::web::Request request{
         .method = method,
         .path = path,
@@ -394,7 +394,7 @@ void OrderEntry::get_exchange_info() {
     };
     auto sequence = download_.sequence();
     connection_(
-        "exchange_info"_sv,
+        "exchange_info"sv,
         request,
         [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
           auto trace_info = server::create_trace_info();
@@ -411,9 +411,9 @@ void OrderEntry::get_exchange_info_ack(
     auto state = OrderEntryState::EXCHANGE_INFO;
     try {
       auto [status, category, body] = response.result();
-      log::debug(R"(status={}, category={}, body="{}")"_sv, status, category, body);
+      log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
       if (download_.skip(sequence, state)) {
-        log::info("Download state={} has already been processed"_sv, state);
+        log::info("Download state={} has already been processed"sv, state);
         return;
       }
       response.expect(core::http::Status::OK);
@@ -423,7 +423,7 @@ void OrderEntry::get_exchange_info_ack(
       (*this)(event);
       download_.check(state);
     } catch (core::NetworkError &e) {
-      log::warn(R"(Exception type={}, what="{}")"_sv, typeid(e).name(), e.what());
+      log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
       download_.retry(state);
     }
   });
@@ -431,13 +431,13 @@ void OrderEntry::get_exchange_info_ack(
 
 void OrderEntry::operator()(const server::Trace<json::ExchangeInfo> &event) {
   auto &[trace_info, exchange_info] = event;
-  log::info<4>("exchange_info={}"_sv, exchange_info);
+  log::info<4>("exchange_info={}"sv, exchange_info);
   std::vector<std::string> symbols;
   size_t counter = {};
   for (const auto &item : exchange_info.symbols) {
-    log::info<2>("item={}"_sv, item);
+    log::info<2>("item={}"sv, item);
     if (shared_.discard_symbol(item.symbol)) {
-      log::info<1>(R"(Drop symbol="{}")"_sv, item.symbol);
+      log::info<1>(R"(Drop symbol="{}")"sv, item.symbol);
       continue;
     }
     // note! convert to lowercase
@@ -483,7 +483,7 @@ void OrderEntry::operator()(const server::Trace<json::ExchangeInfo> &event) {
     };
     create_trace_and_dispatch(handler_, trace_info, market_status, true);
   }
-  log::info("Exchange info: including symbols {}/{}"_sv, counter, exchange_info.symbols.size());
+  log::info("Exchange info: including symbols {}/{}"sv, counter, exchange_info.symbols.size());
   if (!symbols.empty()) {
     SymbolsUpdate symbols_update{
         .symbols = symbols,
@@ -501,7 +501,7 @@ void OrderEntry::new_order(
     if (!ready())
       throw oms::NotReadyException();
     auto method = core::http::Method::POST;
-    auto path = "/api/v3/order"_sv;
+    auto path = "/api/v3/order"sv;
     auto timestamp = core::get_realtime_clock();
     auto side = json::map(create_order.side).as_raw_text();
     auto type = json::map(create_order.order_type).as_raw_text();
@@ -520,7 +520,7 @@ void OrderEntry::new_order(
         R"("icebergQty":{},)"  // XXX ???
         R"("recvWindow":{},)"
         R"("timestamp":{})"
-        R"(}})"_sv,
+        R"(}})"sv,
         create_order.symbol,
         side,
         type,
@@ -534,8 +534,8 @@ void OrderEntry::new_order(
         std::chrono::duration_cast<std::chrono::milliseconds>(Flags::rest_order_recv_window())
             .count(),
         timestamp.count());
-    log::debug(R"(body="{}")"_sv, body);
-    auto headers = fmt::format("X-MBX-APIKEY: {}\r\n"_sv, security_.get_api_key());
+    log::debug(R"(body="{}")"sv, body);
+    auto headers = fmt::format("X-MBX-APIKEY: {}\r\n"sv, security_.get_api_key());
     core::web::Request request{
         .method = method,
         .path = path,
@@ -560,14 +560,14 @@ void OrderEntry::new_order_ack(const server::Trace<core::web::Response> &event) 
     auto &[trace_info, response] = event;
     try {
       auto [status, category, body] = response.result();
-      log::debug(R"(status={}, category={}, body="{}")"_sv, status, category, body);
+      log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
       response.expect(core::http::Status::OK);
       core::json::Buffer buffer(decode_buffer_);
       auto new_order = core::json::Parser::create<json::NewOrder>(body, buffer);
       server::Trace event(trace_info, new_order);
       (*this)(event);
     } catch (core::NetworkError &e) {
-      log::warn(R"(Exception type={}, what="{}")"_sv, typeid(e).name(), e.what());
+      log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
       // XXX HANS ???
     }
   });
@@ -575,7 +575,7 @@ void OrderEntry::new_order_ack(const server::Trace<core::web::Response> &event) 
 
 void OrderEntry::operator()(const server::Trace<json::NewOrder> &event) {
   auto &[trace_info, new_order] = event;
-  log::info<1>("new_order={}"_sv, new_order);
+  log::info<1>("new_order={}"sv, new_order);
   throw NotImplementedException();
 }
 
@@ -590,7 +590,7 @@ void OrderEntry::cancel_order(
     if (!ready())
       throw oms::NotReadyException();
     auto method = core::http::Method::DELETE;
-    auto path = "/api/v3/order"_sv;
+    auto path = "/api/v3/order"sv;
     auto timestamp = core::get_realtime_clock();
     auto body = fmt::format(
         R"({{)"
@@ -599,15 +599,15 @@ void OrderEntry::cancel_order(
         R"("newClientOrderId":"{}")"
         R"("recvWindow":{},)"
         R"("timestamp":{})"
-        R"(}})"_sv,
+        R"(}})"sv,
         order.symbol,
         order.external_order_id,
         request_id,
         std::chrono::duration_cast<std::chrono::milliseconds>(Flags::rest_order_recv_window())
             .count(),
         timestamp.count());
-    log::debug(R"(body="{}")"_sv, body);
-    auto headers = fmt::format("X-MBX-APIKEY: {}\r\n"_sv, security_.get_api_key());
+    log::debug(R"(body="{}")"sv, body);
+    auto headers = fmt::format("X-MBX-APIKEY: {}\r\n"sv, security_.get_api_key());
     core::web::Request request{
         .method = method,
         .path = path,
@@ -632,13 +632,13 @@ void OrderEntry::cancel_order_ack(const server::Trace<core::web::Response> &even
     auto &[trace_info, response] = event;
     try {
       auto [status, category, body] = response.result();
-      log::debug(R"(status={}, category={}, body="{}")"_sv, status, category, body);
+      log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
       response.expect(core::http::Status::OK);
       auto cancel_order = core::json::Parser::create<json::CancelOrder>(body);
       server::Trace event(trace_info, cancel_order);
       (*this)(event);
     } catch (core::NetworkError &e) {
-      log::warn(R"(Exception type={}, what="{}")"_sv, typeid(e).name(), e.what());
+      log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
       // XXX HANS ???
     }
   });
@@ -646,7 +646,7 @@ void OrderEntry::cancel_order_ack(const server::Trace<core::web::Response> &even
 
 void OrderEntry::operator()(const server::Trace<json::CancelOrder> &event) {
   auto &[trace_info, cancel_order] = event;
-  log::info<1>("cancel_order={}"_sv, cancel_order);
+  log::info<1>("cancel_order={}"sv, cancel_order);
   throw NotImplementedException();
 }